+++ SAI/app/Http/Controllers/MunicipioController.php
@@ -6,6 +6,8 @@ use Illuminate\Http\Request;
 use App\Estado;
 use App\Municipio;
 use Laracast\Flash\Flash;
+use App\Http\Requests\MunicipioRequest;
+use Illuminate\Support\Facades\DB;
 
 class MunicipioController extends Controller
 {
@@ -40,15 +42,19 @@ class MunicipioController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(MunicipioRequest $request)
     {
         //dd($request->all());
+        
+        
+        //DB::beginTransaction();    
+            $municipio = new Municipio($request->all());
+            $municipio->save();
+        //DB::commit();    
+            flash("Registro del municipio " .$request->mun_nombre . " exitosamente.")->success();
+            return redirect()->route('municipio.index');
 
-        $municipio = new Municipio($request->all());
-        $municipio->save();
-
-        flash("Registro del municipio " .$request->mun_nombre . " exitosamente.")->success();
-        return redirect()->route('municipio.index');
+       
     }
 
     /**
@@ -60,8 +66,13 @@ class MunicipioController extends Controller
     public function show($id)
     {
         $municipio = Municipio::find($id);
-
-        return view ('admin.lugar.municipio.show',['municipio'=>$municipio]);
+        if ($municipio !== null) {
+           
+            return view ('admin.lugar.municipio.show',['municipio'=>$municipio]);
+        }else{  
+            flash('No hay ningun registro en la Base de Datos del objeto buscado.')->error();
+            return redirect()->route('municipio.index');
+        }
     }
 
     /**
@@ -73,10 +84,17 @@ class MunicipioController extends Controller
     public function edit($id)
     {
         $municipio = Municipio::find($id);
-        $estados = Estado::select('est_nombre','id')->where('id','>',0)->orderby('est_nombre','asc')->pluck('est_nombre','id');
-        //dd($municipio);
 
-        return view('admin.lugar.municipio.edit',['municipio'=>$municipio, 'estados'=>$estados]);
+        if ($municipio !== null) {
+        
+            $estados = Estado::select('est_nombre','id')->where('id','>',0)->orderby('est_nombre','asc')->pluck('est_nombre','id');
+            //dd($municipio);
+
+            return view('admin.lugar.municipio.edit',['municipio'=>$municipio, 'estados'=>$estados]);
+        }else{  
+            flash('No hay ningun registro en la Base de Datos del objeto buscado.')->error();
+            return redirect()->route('municipio.index');
+        }
     }
 
     /**
@@ -86,17 +104,24 @@ class MunicipioController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(MunicipioRequest $request, $id)
     {
         //dd($request->all());
         $municipio = Municipio::find($id);
-        $municipio->mun_nombre = $request->mun_nombre;
-        $municipio->mun_fk_estado = $request->mun_fk_estado;
+        
+        if ($municipio !== null) {
+        
+            $municipio->mun_nombre = $request->mun_nombre;
+            $municipio->mun_fk_estado = $request->mun_fk_estado;
 
-        $municipio->save();
+            $municipio->save();
 
-        flash("Modificación del municipio exitosamente a ".$municipio->mun_nombre )->success();
-        return redirect()->route('municipio.index');
+            flash("Modificación del municipio exitosamente a ".$municipio->mun_nombre )->success();
+            return redirect()->route('municipio.index');
+        }else{  
+            flash('No hay ningun registro en la Base de Datos del objeto buscado.')->error();
+            return redirect()->route('municipio.index');
+        }
 
     }
 
@@ -109,9 +134,15 @@ class MunicipioController extends Controller
     public function destroy($id)
     {
         $municipio = Municipio::find($id);
-        $municipio->delete();
+        
+        if ($municipio !== null) {
+            $municipio->delete();
 
-        flash("Eliminación del municipio " .$municipio->mun_nombre . " exitosamente.")->success();
-        return redirect()->route('municipio.index');
+            flash("Eliminación del municipio " .$municipio->mun_nombre . " exitosamente.")->success();
+            return redirect()->route('municipio.index');
+        }else{  
+            flash('No hay ningun registro en la Base de Datos del objeto buscado.')->error();
+            return redirect()->route('municipio.index');
+        }
     }
 }
